added cucumber first unit test - ready to more scenarios

+++ crypto-wallet/src/main/java/com/genesis/crypto/wallet/coincap/HistoryItem.java
@@ -1,17 +1,18 @@
 package com.genesis.crypto.wallet.coincap;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class HistoryItem {
 
-    private Double priceUsd;
+    private BigDecimal priceUsd;
     private Date time;
 
-    public Double getPriceUsd() {
+    public BigDecimal getPriceUsd() {
         return priceUsd;
     }
 
-    public void setPriceUsd(Double priceUsd) {
+    public void setPriceUsd(BigDecimal priceUsd) {
         this.priceUsd = priceUsd;
     }
